requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0
openai==1.2.0
python-dotenv==1.0.0
jinja2==3.1.2